Fixes Camera destoryPicture() NullPointerException issue

+++ app/src/main/java/nl/hanze/myhealth/Camera.java
@@ -45,6 +45,7 @@ public class Camera {
     }
 
     public static void destroyPicture() {
+        if(mCurrentPhotoPath == null) { return; }
         (new File(mCurrentPhotoPath)).delete();
         mCurrentPhotoPath = null;
     }
